Initial commit
The user sends a request to the government agency with photos of identity documents. Once the government validates the information, it creates an account for the user on the blockchain. Gun manufacturers mint NFTs that records the name and the type of firearms. When a user buys or sells a gun physically, he or she also gets a NFT transfered in or out. NFT serves as a disincentive for people to trade guns illegally because the last person who possesses the token will be held liable if the gun is found in other's possession.

The application is still in progress.